git client is universal
